test: replacing snapshot tests with rtl tests part 15
